refactor: pages > home
