+++ src/components/guides/index.ts
@@ -1,0 +1,4 @@
+import GuideSample from './GuideSample';
+import GuideUsage from './GuideUsage';
+
+export { GuideSample, GuideUsage };
