+++ src/components/landing/index.ts
@@ -1,0 +1,3 @@
+import Landing from './Landing';
+
+export default Landing;
